Rewrite Reflection invocation code, as it wasn't working very well
The Play utility Java utility class made some assumptions about how to
invoke methods in a subclass which didn't work in our situation. In
short, it was taking the arguments passed in and getting the class of
each one, then using those classes when it was looking up the method
to invoke. This was failing for the `all` and `startTransaction`
methods because they took a `Notification` parameter which could be any
of a number of subclasses - Play was looking for a method with the
exact subclass signature, not one with `Notification` as a parameter.